Extracting simulation result
